rename zukanlist class and layout
