--- app/src/main/res/layout/zukanfragment.xml
@@ -1,97 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal">
-
-        <LinearLayout
-            android:id="@+id/ZukanListItme1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:orientation="vertical">
-
-
-            <ImageView
-                android:id="@+id/image1"
-                android:layout_width="150dp"
-                android:layout_height="130dp"
-                android:layout_margin="10dp"
-                android:layout_weight="0.41"
-                android:src="@drawable/sakana2"/>
-
-            <TextView
-                android:id="@+id/title1"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="bottom"
-                android:gravity="center"
-                android:text="くらげ"
-                android:textColor="#b7c5cc"
-                android:textSize="24dp"/>
-
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/ZukanListItme2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:orientation="vertical">
-
-            <ImageView
-                android:id="@+id/image2"
-                android:layout_width="150dp"
-                android:layout_height="130dp"
-                android:layout_gravity="center_vertical"
-                android:layout_margin="10dp"
-                android:layout_weight="0.36"/>
-
-
-            <TextView
-                android:id="@+id/title2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center"
-                android:textColor="#b7c5cc"
-                android:textSize="24dp"
-
-                />
-
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/ZukanListItme3"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:layout_margin="10dp"
-            android:orientation="vertical">
-
-            <ImageView
-                android:id="@+id/image3"
-                android:layout_width="150dp"
-                android:layout_height="130dp"
-                android:layout_margin="10dp"
-                android:layout_weight="0.36"/>
-
-            <TextView
-                android:id="@+id/title3"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center"
-                android:textColor="#b7c5cc"
-                android:textSize="24dp"
-                />
-
-        </LinearLayout>
-
-    </LinearLayout>
-</LinearLayout>
